+++ src/Components/WeatherForecast/WeatherForecast.jsx
@@ -1,15 +1,7 @@
 import React from 'react';
+import utcToHHMM from "../../Services/timeConvert"
 
 const WeatherForecast = ({ locationName, weatherForecast }) => {
-  const utcToHHMM = (time) => {
-    const utcTime = new Date(time * 1e3);
-    const addZeroBefore = (n) => {
-      return (n < 10 ? '0' : '') + n;
-    };
-    const currentForecastHour = addZeroBefore(utcTime.getHours());
-    const currentForecastMinutes = addZeroBefore(utcTime.getMinutes());
-    return `${currentForecastHour}:${currentForecastMinutes}`;
-  };
 
   return (
     <section className="p-4 flex w-auto">
